@@ -32,10 +32,10 @@
         ]
     },
     "require": {
-        "php": ">=7.2"
+        "php": ">=7.4"
     },
     "require-dev": {
-        "phpunit/phpunit": "~8.0",
+        "phpunit/phpunit": "~9",
         "squizlabs/php_codesniffer": "3.*",
         "phpstan/phpstan": "^0.12"
     }
